different end_gaps for left and right, f_res calc for meandered
updated the code to allow different end gaps for the left and right
resonators. This is so that if the fingers are bigger you can also
have a bigger or smaller end gap. I want to make big end gaps often
to make the capacitance from that region more negligable.
F_res for meandered polkids was not correct and also need modifications
to account for the process bias of the aluminum wet etchxsx